+++ config/routes.php
@@ -1,3 +1,10 @@
-<?php
+<?php defined('BASEPATH') or exit('No direct script access allowed');
 
+/**
+ * Streams Import Routes
+ *
+ * @package  PyroCMS\Addons\Modules\Streams Import\Config
+ * @author   PyroCMS Community
+ * @website  https://github.com/bergeo-fr/streams_import
+ */
 $route['streams_import/admin/profiles(/:any)?'] = 'admin_profiles$1';