+++ language/french/streams_import_lang.php
@@ -1,26 +1,34 @@
 <?php defined('BASEPATH') or exit('No direct script access allowed');
+
+/**
+ * Streams Import Language
+ *
+ * @package  PyroCMS\Addons\Modules\Streams Import\Lang
+ * @author   PyroCMS Community
+ * @website  https://github.com/bergeo-fr/streams_import
+ */
 $lang = array(
-	//Title
+	# titles
 	'streams_import:title:profiles:index'                   => 'Profils',
 	'streams_import:title:profiles:create'                  => 'Creation d\'un profil',
 	'streams_import:title:profiles:edit'                    => 'Edition d\'un profil',
+	'streams_import:title:profiles:quick_import'            => 'Quick Import', #translate
+	'streams_import:title:profiles:quick_import_mapping'    => 'Quick Import: Mapping', #translate
+	'streams_import:title:profiles:quick_import_success'    => 'Quick Import: Success', #translate
 
-	//Message
+	# messages
 	'streams_import:messages:profiles:create:success'       => 'Nouveau profil créé avec succès.',
 	'streams_import:messages:profiles:create:failure'       => 'Echec dans la création du profil.',
 	'streams_import:messages:profiles:edit:success'         => 'Profil édité avec succès.',
 	'streams_import:messages:profiles:edit:failure'         => 'Echec dans l\'édition du profil.',
 	'streams_import:messages:profiles:delete:success'       => 'Profil supprimé avec succès.',
 	'streams_import:messages:profiles:delete:failure'       => 'Echec dans la suppression du profil.',
-	'streams_import:messages:import:success'                => 'Import Done. Yeah.',
-	'streams_import:messages:import:failure'                => 'Import fail. Sorry about that :(',
-
-
+	'streams_import:messages:import:success'                => 'Import Done. Yeah.', #translate
+	'streams_import:messages:import:failure'                => 'Import fail. Sorry about that :(', #translate
 	'streams_import:messages:mapping:save:success'          => 'Mapping sauvegardé',
 	'streams_import:messages:mapping:save:failure'          => 'Echec du mapping',
 
-	//Fields
-	//customer
+	# fields
 	'streams_import:fields:profile_name'                    => 'Nom du profil',
 	'streams_import:fields:profile'                         => 'Profile ID', #translate
 	'streams_import:fields:entry_number'                    => 'Entrée',
@@ -41,8 +49,7 @@ $lang = array(
 	'streams_import:fields:source_format'                   => 'Source Format', #translate
 	'streams_import:fields:source_format_instructions'      => 'The format of the source data', #translate
 
-
-	//button
+	# buttons
 	'profiles:button:add'                                   => 'Ajouter un profil',
 	'streams_import:button:quick_import'                    => 'Quick Import', #translate
 	'streams_import:button:edit-mapping'                    => 'Editer le mapping',
@@ -51,11 +58,10 @@ $lang = array(
 	'streams_import:button:run'                             => 'Executer',
 	'streams_import:button_next'                            => 'Suivant',
 
-
+	# misc.
 	'streams_import:title_import_csv'                       => 'Importer un fichier',
 	'streams_import:misc_example_csv'                       => 'Un fichier XML / TXT ou XML',
 	'streams_import:misc_csv_file'                          => 'Misc.', #translate
 	'streams_import:misc_instructions_csv_file'             => 'Instructions', #translate
-
 
 );