+++ CHANGELOG.md
@@ -5,9 +5,9 @@ __NOTE:__ All versions prior to _1.0_ are beta.
 
 ## v0.2 / In Development
 - Added "Quick Import" feature to allow import without a Profile
-- Added support for JSON source format
-- Added support for importing by URL instead of file
-- Added support for Streams Core Columns: `id, created, updated, created_by, ordering_count`
+  - Added support for JSON source format
+  - Added support for importing by URL instead of file
+  - Added support for Streams Core Columns: `id, created, updated, created_by, ordering_count`
 
 ## v0.1
 Default functionality working